requirements.txt
# Core
torch>=2.2
torchvision>=0.17
numpy>=1.26
pydantic>=2.7
pyyaml>=6.0.1


# NAS / Optimization
optuna>=3.6


# Utilities
tqdm>=4.66
rich>=13.7


# FLOPs & Carbon
thop>=0.1.1.post2209072238
codecarbon>=2.4 # optional; disable if issues

# Agents Orchestration
crewai
litellm
litellm[proxy]

# Memory
faiss-cpu>=1.8.0 # optional; memory module falls back to numpy if unavailable
